module-8/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
websockets>=11.0
langgraph>=0.0.50
llama-index>=0.9.0
qdrant-client>=1.6.0
openai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.24.0
python-multipart>=0.0.6
PyPDF2>=3.0.0
sentence-transformers>=2.2.0
torch>=1.9.0

# LangChain dependencies for document loaders
langchain>=0.1.0
langchain-community>=0.0.10
langchain-text-splitters>=0.0.1
unstructured>=0.10.0
pdfplumber>=0.9.0
python-docx>=0.8.11
openpyxl>=3.1.0

# Visualization dependencies
graphviz>=0.20.0  # For PNG graph visualization
ipython>=8.0.0    # For Jupyter display functionality
grandalf>=0.8     # For ASCII graph visualization 